Restructure visualization utilities for maintainability
- Split `utilities/data_visualization.py` into three new modules:
  - `utilities/visualization/control_plot.py`
  - `utilities/visualization/control_plot_anim.py`
  - `utilities/visualization/plot_utilities.py`
- Update unit tests to align with the restructured visualization
  modules.
- Update `README.md` to reflect the visualization utilities
  reorganization.

This change improves maintainability by separating plotting
utilities, ensuring clear separation of concerns. It also reduces
cluttering and improves code organization.
