+++ tests/test_nonlinear_dd_mpc_integration.py
@@ -21,13 +21,15 @@ from direct_data_driven_mpc.utilities.controller.data_driven_mpc_sim import (
 from direct_data_driven_mpc.utilities.controller.initial_data_generation import (  # noqa: E501
     generate_initial_input_output_data,
 )
-from direct_data_driven_mpc.utilities.data_visualization import (
-    plot_input_output,
-    plot_input_output_animation,
-    save_animation,
-)
 from direct_data_driven_mpc.utilities.models.nonlinear_model import (
     NonlinearSystem,
+)
+from direct_data_driven_mpc.utilities.visualization.control_plot import (
+    plot_input_output,
+)
+from direct_data_driven_mpc.utilities.visualization.control_plot_anim import (
+    plot_input_output_animation,
+    save_animation,
 )
 
 matplotlib.use("Agg")  # Prevent GUI backend
